Move tests module to tests.rs file
